requirements.txt
boto3==1.35.73
botocore==1.35.73
gTTS==2.5.3
moviepy==1.0.3
playwright==1.48.0
praw==7.8.1
prawcore~=2.4.0
requests==2.32.3
rich==13.9.4
toml==0.10.2
translators==5.9.3
pyttsx3==2.98
Pillow>=10.4.0
tomlkit==0.13.2
Flask==3.1.0
clean-text==0.6.0
unidecode==1.3.8
spacy==3.8.2
torch>=2.4.0
transformers==4.47.1
ffmpeg-python==0.2.0
elevenlabs==1.15.0
yt-dlp>=2024.12.3
numpy>=1.26.4
